requirements.txt
gunicor
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-WTF>=1.2
Flask-Login>=0.6
Flask-Bcrypt>=1.0
Flask-Migrate>=4.0
Flask-Limiter>=3.8
Flask-Mail>=0.9.1
python-dotenv>=1.0
redis>=5
